Possibility to dump CSV files from the statistics (still needs refactoring)
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/muauto@17960 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/statistics/StatsTest.java
@@ -134,8 +134,8 @@ public class StatsTest {
         "Property 2",
         Aggregateables.AggregationInt.class);
 
-    assertEquals(787, p1stats.getSum());
-    assertEquals(10, p2stats.getSum());
+    assertEquals(787, (int) p1stats.getSum());
+    assertEquals(10, (int) p2stats.getSum());
     assertEquals(2, p1stats.getValuations());
     assertEquals(1, p2stats.getValuations());
   }
